+++ changelog.md
@@ -2,6 +2,7 @@
 
 - Add `is_iterable`.
 - Add `WeakSet`.
+- Add `Future.wait_for_completion`.
 
 ## 1.0.8 *\[2022-01-04\]*